refactor: 약속 장소 및 시간 ui model 을 사용하여 관리
