+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/commentdetail/adapter/participant/ParticipantViewHolder.kt
@@ -2,7 +2,7 @@ package com.zzang.chongdae.presentation.view.commentdetail.adapter.participant
 
 import androidx.recyclerview.widget.RecyclerView
 import com.zzang.chongdae.databinding.ItemOfferingMemberBinding
-import com.zzang.chongdae.presentation.view.commentdetail.model.ParticipantUiModel
+import com.zzang.chongdae.presentation.view.commentdetail.model.participants.ParticipantUiModel
 
 class ParticipantViewHolder(
     private val binding: ItemOfferingMemberBinding,
