--- android/app/src/main/java/com/zzang/chongdae/presentation/view/commentdetail/model/ParticipantUiModel.kt
@@ -1,5 +1,0 @@
-package com.zzang.chongdae.presentation.view.commentdetail.model
-
-data class ParticipantUiModel(
-    val nickname: String,
-)
