--- android/app/src/main/java/com/zzang/chongdae/presentation/view/commentdetail/model/ProposerUiModel.kt
@@ -1,5 +1,0 @@
-package com.zzang.chongdae.presentation.view.commentdetail.model
-
-data class ProposerUiModel(
-    val nickname: String,
-)
