+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/commentdetail/model/meeting/MeetingsUiModel.kt
@@ -1,0 +1,9 @@
+package com.zzang.chongdae.presentation.view.commentdetail.model.meeting
+
+import java.time.LocalDateTime
+
+class MeetingsUiModel(
+    val meetingDate: LocalDateTime,
+    val meetingAddress: String,
+    val meetingAddressDetail: String?,
+)
