+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/commentdetail/model/participants/ParticipantUiModel.kt
@@ -1,0 +1,5 @@
+package com.zzang.chongdae.presentation.view.commentdetail.model.participants
+
+data class ParticipantUiModel(
+    val nickname: String,
+)
