+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/commentdetail/model/participants/ProposerUiModel.kt
@@ -1,0 +1,5 @@
+package com.zzang.chongdae.presentation.view.commentdetail.model.participants
+
+data class ProposerUiModel(
+    val nickname: String,
+)
