+++ android/app/src/main/res/layout/activity_comment_detail.xml
@@ -88,7 +88,6 @@
                 android:layout_width="285dp"
                 android:layout_height="@dimen/size_100"
                 android:layout_marginTop="@dimen/margin_20"
-                android:background="@drawable/bg_white_radius_16dp"
                 android:contentDescription="@string/comment_detail"
                 android:scaleType="centerCrop"
                 app:layout_constraintEnd_toEndOf="parent"
@@ -191,7 +190,7 @@
                     android:layout_marginStart="@dimen/margin_10"
                     android:ellipsize="end"
                     android:maxLines="2"
-                    android:text="@{vm.location}"
+                    android:text="@{vm.meetings.meetingAddress}"
                     android:textColor="@color/gray_font"
                     android:textSize="@dimen/size_11"
                     app:isVisible="@{vm.isCollapsibleViewVisible}"
@@ -205,7 +204,7 @@
                     android:layout_height="wrap_content"
                     android:ellipsize="end"
                     android:maxLines="2"
-                    android:text="@{vm.locationDetail}"
+                    android:text="@{vm.meetings.meetingAddressDetail}"
                     android:textColor="@color/gray_font"
                     android:textSize="@dimen/size_11"
                     app:isVisible="@{vm.isCollapsibleViewVisible}"
@@ -233,7 +232,7 @@
                     android:maxLines="1"
                     android:textColor="@color/gray_font"
                     android:textSize="@dimen/size_11"
-                    app:formattedDeadline="@{vm.deadline}"
+                    app:formattedDeadline="@{vm.meetings.meetingDate}"
                     app:isVisible="@{vm.isCollapsibleViewVisible}"
                     app:layout_constraintBottom_toBottomOf="@id/iv_calendar"
                     app:layout_constraintStart_toEndOf="@id/iv_calendar"
