+++ android/app/src/main/res/layout/item_offering_member.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="participant"
-            type="com.zzang.chongdae.presentation.view.commentdetail.model.ParticipantUiModel" />
+            type="com.zzang.chongdae.presentation.view.commentdetail.model.participants.ParticipantUiModel" />
     </data>
